Move Entitis and repositories their specific directories
