+++ reader/src/main/java/com/softserveinc/ita/homeproject/reader/data/entity/general/news/News.java
@@ -1,8 +1,8 @@
-package com.softserveinc.ita.homeproject.reader.data.general.news;
+package com.softserveinc.ita.homeproject.reader.data.entity.general.news;
 
 import java.time.LocalDateTime;
 
-import com.softserveinc.ita.homeproject.reader.data.BaseEntity;
+import com.softserveinc.ita.homeproject.reader.data.entity.BaseEntity;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import org.springframework.data.mongodb.core.mapping.Document;
